* pcb/impy: (shield.sch) Change whole AD5933 external analog front end.
The new front end uses 4-wire measurement to eliminate the series resistance of the analog multiplexers and test leads.
Using four 16:1 multiplexers, up to 10 channels can be used together with 6 calibration resistors.
The output signal from the AD5933 is AC coupled so as to eliminate the DC bias, which limits the dynamic range.
With the option of four selectable attenuation values for the AD5933 output voltage, impedances down to about 10 Ohms can be measured, if the chip is also clocked from a slow external source the measurement frequency can be as low as 10 Hz.