* stm/impy: (ad5933.c) Add charging of the coupling capacitor to AD5933_StartMeasurement function, we always need to do this, assuming the output was off before.
This needs some work, charging the capacitor takes a significant amount of time. If autoranging is implemented outside the AD5933 driver then we can expect many measurements to be started until the right range is found, which would take insane amounts of time. Either autoranging is integrated into the driver or we need a means of ensuring that the output does not get switched off and the capacitor stays charged to the right voltage.