* stm/impy: (usbd_vcp_if.c) This is a rather massive commit that changes the way data transmission and the command line are handled, so let's dive right in.
The structure of the transmission functions has changed in that they don't actually send data any more, but rather copy it to the transmit buffer. The new VCP_Flush function needs to be called to actually transmit the data. This fixes some ugliness when multiple strings need to be sent consecutively.
The command line now sends a CR-LF combo regardless of what line termination is received, mainly to be compatible with the standard settings of PuTTY (and because I hope that other terminal software accepts this as well).
Furthermore, processing of received characters is stopped after a line break is received (only one command at a time is processed).
An error is fixed in the SendBuffer function where a scheduled transmission of an external buffer is ignored.
And of course the documentation is updated to reflect the changes.