* stm/impy: (ad5933.c) Implement clock source change during a sweep.
The phase error when using a different clock source than was calibrated with is considerable, maybe we need to do a phase calibration in each frequency range. Therefore the gain factor structures would need to be changed and it would be nice if the calibration process and usage of the gain factor data would remain transparent to the user of the driver.
The alternative would be to let the user manage multiple gain factors for different frequency ranges.